backend/requirements.txt
numpy
pillow
pymongo
urllib3==1.25.3
